feat: remove requestExit and add multicall
