+++ src/plugin.h
@@ -30,9 +30,9 @@
     X(BATCH_DEPOSIT, 0xc82655b7)             \
     X(BATCH_COLLECT_REWARD, 0x3ec0c7b5)      \
     X(MINT, 0xa0712d68)                      \
-    X(REQUEST_EXIT, 0x7f8e3b4e)              \
     X(COLLECT_REWARD, 0x1bc1f6e1)            \
-    X(COLLECT_REWARD_FOR_NFT, 0x58f2b964)
+    X(COLLECT_REWARD_FOR_NFT, 0x58f2b964)    \
+    X(MULTICALL, 0xbfe6646e)
 
 // Xmacro helpers to define the enum and map
 // Do not modify !
