--- tests/abis/0xC57beAe054B45Bb49B65b42C49c6f4b010f976e5.abi.json
@@ -1,9 +1,0 @@
-[
-    {
-        "inputs": [],
-        "name": "requestExit",
-        "outputs": [],
-        "stateMutability": "nonpayable",
-        "type": "function"
-    }
-]
