+++ tests/snapshots/nanox/test_multicall/00000.png
