+++ tests/snapshots/nanox/test_multicall/00003.png
