+++ tests/snapshots/nanox/test_multicall/00004.png
