+++ tests/snapshots/nanox/test_multicall/00005.png
