--- tests/test_mint.py
@@ -1,70 +1,0 @@
-from pathlib import Path
-import json
-import os
-
-import datetime
-
-from web3 import Web3
-from eth_typing import ChainId
-
-from ledger_app_clients.ethereum.client import EthAppClient, StatusWord
-from ledger_app_clients.ethereum.utils import get_selector_from_data
-from ragger.navigator import NavInsID
-
-from .utils import get_appname_from_makefile
-
-
-ROOT_SCREENSHOT_PATH = Path(__file__).parent
-ABIS_FOLDER = "%s/abis" % (os.path.dirname(__file__))
-
-PLUGIN_NAME = get_appname_from_makefile()
-
-with open("%s/0x5B41FFB9C448C02Ff3D0401b0374b67EFcB73C7E.abi.json" % (ABIS_FOLDER)) as file:
-    contract = Web3().eth.contract(
-        abi=json.load(file),
-        # Get address from filename
-        address=bytes.fromhex(os.path.basename(file.name).split(".")[0].split("x")[-1])
-    )
-
-def test_mint(backend, firmware, navigator, test_name):
-    client = EthAppClient(backend)
-
-    # values taken from https://goerli.etherscan.io/tx/0x9dbba686f9cd1911ec41fead1b2de135ee2eea8f166c07a25326d79775219aa0
-    data = contract.encodeABI(
-        fn_name="mint",
-        args=[2],
-    )
-
-    # first setup the external plugin
-    with client.set_external_plugin(PLUGIN_NAME,
-                                    contract.address,
-                                    # Extract function selector from the encoded data
-                                    get_selector_from_data(data)):
-        pass
-
-    # send the transaction
-    with client.sign("m/44'/60'/1'/0/0", {
-             "nonce": 20,
-             "maxFeePerGas": Web3.to_wei(145, "gwei"),
-             "maxPriorityFeePerGas": Web3.to_wei(1.5, "gwei"),
-             "gas": 173290,
-             "to": contract.address,
-             "value": Web3.to_wei(32, "ether"),
-             "chainId": ChainId.GOR,
-             "data": data
-         }):
-
-        # Validate the on-screen request by performing the navigation appropriate for this device
-        if firmware.device.startswith("nano"):
-            navigator.navigate_until_text_and_compare(NavInsID.RIGHT_CLICK,
-                                                      [NavInsID.BOTH_CLICK],
-                                                      "Accept",
-                                                      ROOT_SCREENSHOT_PATH,
-                                                      test_name)
-        else:
-            navigator.navigate_until_text_and_compare(NavInsID.USE_CASE_REVIEW_TAP,
-                                                      [NavInsID.USE_CASE_REVIEW_CONFIRM,
-                                                       NavInsID.USE_CASE_STATUS_DISMISS],
-                                                      "Hold to sign",
-                                                      ROOT_SCREENSHOT_PATH,
-                                                      test_name)
