+++ tests/test_nfts.py
@@ -19,20 +19,66 @@ ABIS_FOLDER = "%s/abis" % (os.path.dirname(__file__))
 
 PLUGIN_NAME = get_appname_from_makefile()
 
-with open("%s/0xC57beAe054B45Bb49B65b42C49c6f4b010f976e5.abi.json" % (ABIS_FOLDER)) as file:
+with open("%s/0x5B41FFB9C448C02Ff3D0401b0374b67EFcB73C7E.abi.json" % (ABIS_FOLDER)) as file:
     contract = Web3().eth.contract(
         abi=json.load(file),
         # Get address from filename
         address=bytes.fromhex(os.path.basename(file.name).split(".")[0].split("x")[-1])
     )
 
-def test_request_exit(backend, firmware, navigator, test_name):
+def test_mint(backend, firmware, navigator, test_name):
     client = EthAppClient(backend)
 
     # values taken from https://goerli.etherscan.io/tx/0x9dbba686f9cd1911ec41fead1b2de135ee2eea8f166c07a25326d79775219aa0
     data = contract.encodeABI(
-        fn_name="requestExit",
-        args=[],
+        fn_name="mint",
+        args=[2],
+    )
+
+    # first setup the external plugin
+    with client.set_external_plugin(PLUGIN_NAME,
+                                    contract.address,
+                                    # Extract function selector from the encoded data
+                                    get_selector_from_data(data)):
+        pass
+
+    # send the transaction
+    with client.sign("m/44'/60'/1'/0/0", {
+             "nonce": 20,
+             "maxFeePerGas": Web3.to_wei(145, "gwei"),
+             "maxPriorityFeePerGas": Web3.to_wei(1.5, "gwei"),
+             "gas": 173290,
+             "to": contract.address,
+             "value": Web3.to_wei(32, "ether"),
+             "chainId": ChainId.GOR,
+             "data": data
+         }):
+
+        # Validate the on-screen request by performing the navigation appropriate for this device
+        if firmware.device.startswith("nano"):
+            navigator.navigate_until_text_and_compare(NavInsID.RIGHT_CLICK,
+                                                      [NavInsID.BOTH_CLICK],
+                                                      "Accept",
+                                                      ROOT_SCREENSHOT_PATH,
+                                                      test_name)
+        else:
+            navigator.navigate_until_text_and_compare(NavInsID.USE_CASE_REVIEW_TAP,
+                                                      [NavInsID.USE_CASE_REVIEW_CONFIRM,
+                                                       NavInsID.USE_CASE_STATUS_DISMISS],
+                                                      "Hold to sign",
+                                                      ROOT_SCREENSHOT_PATH,
+                                                      test_name)
+
+def test_multicall(backend, firmware, navigator, test_name):
+    client = EthAppClient(backend)
+
+    # values taken from https://goerli.etherscan.io/tx/0x9dbba686f9cd1911ec41fead1b2de135ee2eea8f166c07a25326d79775219aa0
+    data = contract.encodeABI(
+        fn_name="multicall",
+        args=[
+            [221],
+            ["0x3ccfd60b"]
+        ],
     )
 
     # first setup the external plugin
@@ -50,7 +96,7 @@ def test_request_exit(backend, firmware, navigator, test_name):
              "gas": 173290,
              "to": contract.address,
              "value": Web3.to_wei(0, "ether"),
-             "chainId": ChainId.GOR,
+             "chainId": ChainId.ETH,
              "data": data
          }):
